refactoring testing files, will still need to fix any tests that fails
