+++ src/test/UseCasesTest/Comparator/ProductComparatorFactoryTest.java
@@ -1,3 +1,5 @@
+package UseCasesTest.Comparator;
+
 import Entities.*;
 import UseCases.Comparator.ProductComparator;
 import UseCases.Comparator.ProductComparatorFactory;
